+++ src/main/java/warmup/one/WarmupOneFifthSet.java
@@ -25,4 +25,21 @@ public class WarmupOneFifthSet {
     public boolean or35(int n) {
         return n % 3 == 0 || n % 5 == 0;
     }
+
+    /**
+     * Given a string, take the first 2 chars and return the string with the 2 chars added at both
+     * the front and back, so "kitten" yields "kikittenki". If the string length is less than 2, use
+     * whatever chars are there.
+     *
+     * @param str
+     * @return String
+     */
+    public String front22(String str) {
+        if (str.length() < 2) {
+            return str + str + str;
+        }
+        String firstTwoChars = str.substring(0, 2);
+
+        return firstTwoChars + str + firstTwoChars;
+    }
 }