+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Warmup-1 front3
 - Warmup-1 backAround
 - Warmup-1 or35
+- Warmup-1 front22
 
 ## [1.0.0] - 2022-09-24
 ### Added